requirements.txt
# ── Telegram Bot ────────────────────────────────────────────────────────────
aiogram==3.7.0
python-dotenv>=1.0.0

# ── Async Database ───────────────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Finance / MAC3 Risk Engine ───────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.36
scikit-learn>=1.4.0
requests>=2.31.0
edgartools>=3.0.0

# ── Security / Encryption ────────────────────────────────────────────────────
cryptography>=42.0.0

# ── PDF Generation ───────────────────────────────────────────────────────────
jinja2>=3.1.3
playwright>=1.44.0

# ── AI (advisor_bot / rag_engine) ────────────────────────────────────────────
anthropic>=0.25.0
chromadb>=0.5.0
pymupdf4llm>=0.0.10
